play around with a more generalized mechanism for contextual schemas
this is a way that allows you to dynamically choose a schema in a more general way than when_key_is and when_key_exists.

The downside is that it is not strictly data-driven, like when_key_is and when_key_exists -- it requires putting Python functions in your schemas.