add a new `when_type_is` directive to the `choose_schema` directive
this allows choosing a schema based on the type.

With this change, I believe that use of the `anyof` and `oneof` directives should basically be unheard of.

At least in my (fairly extensive) usage of sureberus, the `when_type_is` directive will replace every last use of `anyof` and `oneof`.